requirements.txt
fastapi
fastapi-sqlalchemy
pydantic
uvicorn
alembic
psycopg2-binary
python-dotenv